# docs/static/swagger/swagger.yaml
openapi: 3.0.3
info:
  title: eInvoice Storage - OpenAPI 3.0
  description: |-
    A description of the Invoicer REST API
  contact:
    email: patricklachlanbell@gmail.com
  version: 1.0.1
...
servers:
  - url: http://node-express-env.eba-9atmj7qr.ap-southeast-2.elasticbeanstalk.com
    description: Production server (uses live data and can be viewed on frontend)
paths:
  /invoice:
    post:
      tags:
        - invoice
      summary: Create a new invoice
      description: Uses the provided metadata and contents to create a new invoice for the authenticated user. Returns the new invoices id.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - documentTitle
                - totalInvoiceValue
                - supplierName
                - recipientName
                - contentsXml
                - jsonXml
              description: Only one of contentsXml and jsonXml is required. The one which is missing will be generated from the other.
              properties:
                contentsXml:
                  example: "<Invoice></Invoice>"
                  type: string
                  minLength: 1
                contentsJson:
                  type: string
                  minLength: 1
                documentTitle:
                  example: "Example"
                  type: string
                totalInvoiceValue:
                  example: 1000
                  description: Currency is assumed to be AUD.
                  type: number
                supplierName:
                  example: ""
                  type: string
                recipientName:
                  example: "Company A LLC"
                  type: string
      responses:
        '200':
          description: Successfully inserted an invoice
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                  - insertedId
                properties:
                  acknowledged:
                    type: boolean
                  insertedId:
                    type: string
                    minLength: 24
                    maxLength: 24
                    example: "6421d1e887d1e2174b494492"
        '400':
          $ref: '#/components/responses/InvalidError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          description: Internal server error
      security:
        - bearerAuth: []
  /invoice/{id}:
    get:
      tags:
        - invoice
      summary: Retrieve the invoice with the specified id.
      description: The specified id is searched for. If it is found and the authorizing user can view the invoice, the contents and metadata are returned.
      parameters:
        - name: id
          in: path
          description: The objectId of the invoice to be retrieved.
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      responses:
        '200':
          description: Found a viewable invoice with the given id.
          content:
            application/json:
              schema:
                type: object
                required:
                  - found
                properties:
                  found:
                    type: boolean
                    example: true
                  invoice:
                    $ref: '#/components/schemas/Invoice'
        '400':
          $ref: '#/components/responses/InvalidError'
        '404':
          description: No invoice with the id was found or user is not authorized to view this invoice.
          content:
            application/json:
              schema:
                type: object
                required:
                  - found
                properties:
                  found:
                    type: boolean
                    example: false
        '500':
          description: Internal server error
      security:
        - bearerAuth:
            - read:invoice
    delete:
      tags:
        - invoice
      summary: Delete the invoice with the specified id.
      description: The invoice with the given id is deleted if it is found and the authorizing user has edit permissions.
      parameters:
        - name: id
          in: path
          description: The objectId of invoice to be deleted.
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      responses:
        '200':
          description: Invoice was found and deleted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - deletedCount
                  - acknowledged
                properties:
                  acknowledged:
                    type: boolean
                    example: true
                  deletedCount:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/InvalidError'
        '404':
          description: An invoice with id doesn't exist or user is not authorized to delete it.
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                  - deletedCount
                properties:
                  acknowledged: 
                    type: boolean
                    example: true
                  deletedCount:
                    type: integer
                    example: 0
        '500':
          description: Internal server error
      security:
        - bearerAuth:
            - edit:invoice
    patch:
      tags:
        - invoice
      summary: Modify the invoice with the specified id.
      description: An invoice with the specified id is searched for and if it is found the contents are modified according to the request body. All fields in the body must be valid possible properties of an invoice.
      parameters:
        - name: id
          in: path
          description: objectId of invoice to be modified.
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Invoice'
        required: false
      responses:
        '200':
          description: The invoice with the given id was successfully modified.
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                properties:
                  aknowledged:
                    type: boolean
                    example: true
                  matchedCount:
                    type: integer
                    example: 1
        '400':
          $ref: '#/components/responses/InvalidError'
        '404':
          description: An invoice with the given id doesn't exist or the user is not authorized to edit it.
          content:
            application/json:
              schema:
                type: object
                required:
                  - acknowledged
                properties:
                  aknowledged:
                    type: boolean
                    example: true
                  matchedCount:
                    type: integer
                    example: 0
        '500':
          description: Internal server error
      security:
        - bearerAuth:
            - read:invoice
  /invoice/view/{id}:
    get:
      tags:
        - invoice
      summary: View an invoice.
      description: Returns a HTML page which shows the details of an invoice stored in the database.
      parameters:
        - name: id
          in: path
          description: The objectId of the invoice to be viewed.
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      responses:
        '200':
          description: Generated HTML for the invoice.
          content:
            application/json:
              schema:
                type: object
                required:
                  - found
                  - html
                properties:
                  found:
                    type: boolean
                    example: true
                  html:
                    type: string
                    example: "<html>Invoice</html>"
                    description: Generated HTML from the invoice's XML or JSON representation. This expects the data to be formatted according to the PEPPOL regulations. JSON should be a direct transposition of this PEPPOL structure, for reference, we use fast-xml-parser with ingnoreAttributes off, and implicitly expect the JSON to follow the output of this parser.
        '404':
          description: The invoice was not found or the authenticating user does not have view permissions.
          content:
            application/json:
              schema:
                type: object
                required:
                  - found
                properties:
                  found:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "No matching invoice found"
        '400':
          $ref: '#/components/responses/InvalidError'
        '500':
          description: Internal server error.
      security:
        - bearerAuth:
          - view:invoice
  /invoice/view:
    post:
      tags:
        - invoice
      summary: View an invoice.
      description: Returns a HTML page which shows the details of an invoice from it's XML or JSON contents.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: One of contentsXml or contentsJson is required. If both are provided, the JSON is used.
              properties:
                contentsXml:
                  type: string
                  description: For the expected document schema check the PEPPOL standard. An example of a valid XML string is shown here https://github.com/A-NZ-PEPPOL/A-NZ-PEPPOL-BIS-3.0/blob/master/Message%20examples/AU%20Freight%20-%20Document%20Level.xml
                contentsJson:
                  type: string
                  description: The JSON form of a PEPPOL standard XML, the JSON structure should directly follows the XML. For reference, we use fast-xml-parser with ingnoreAttributes off to convert from the XML, and implicitly expect the JSON to follow the output of this parser.
        required: true
      responses:
        '200':
          description: Generated HTML for the invoice.
          content:
            application/json:
              schema:
                type: object
                required:
                  - found
                  - html
                properties:
                  found:
                    type: boolean
                    example: true
                  html:
                    type: string
                    example: "<html>Invoice</html>"
                    description: Generated HTML from the invoice's XML or JSON representation.
        '400':
          $ref: '#/components/responses/InvalidError'
        '500':
          description: Internal server error.
  /invoice/send/{id}:
    post:
      tags:
        - invoice
      summary: Send an email containing an invoice.
      description: Sends an email to a list of provided recipients providing information about an invoice.
      parameters:
        - name: id
          in: path
          description: The objectId of the invoice to be sent.
          required: true
          schema:
            type: string
            minLength: 24
            maxLength: 24
            example: "6421d1e887d1e2174b494492"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - recipients
              properties:
                recipients:
                  type: array
                  description: A list of emails addresses which will receive the email.
                  example: ["johndoe@example.com"]
                  items:
                    type: string
                    example: "johndoe@example.com"
                    description: Must be a correctly formatted email address.
        required: true
      responses:
        '200':
          description: Email was succesfully sent. Note that this does not mean the email will show up in the recipient's inbox. The email may take a long time to arrive, may be reported as spam, or even blocked entirely.
          content:
            application/json:
              schema:
                type: object
                required:
                  - sent
                properties:
                  found:
                    type: boolean
                    example: true
        '404':
          description: The invoice was not found or the authenticating user does not have view permissions.
          content:
            application/json:
              schema:
                type: object
                required:
                  - sent
                properties:
                  sent:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "No matching invoice found"
        '400':
          $ref: '#/components/responses/InvalidError'
        '500':
          description: Internal server error.
      security:
        - bearerAuth:
          - view:invoice
...
components:
  schemas:
...
    Invoice:
      type: object
      required:
        - documentTitle
        - totalInvoiceValue
        - supplierName
        - recipientName
        - viewIds
        - editIds
        - creationTime
        - creatorId
      properties:
        contentsXml:
          example: "<Invoice></Invoice>"
          type: string
          minLength: 1
        contentsJson:
          type: string
          minLength: 1
        documentTitle:
          example: "Example"
          type: string
        totalInvoiceValue:
          example: 1000
          description: Currency is assumed to be AUD.
          type: number
        supplierName:
          example: "Company A LLC"
          type: string
        recipientName:
          example: "Company B LLC"
          type: string
        viewIds:
          example: ["6421d1e887d1e2174b494492"]
          type: array
          items:
            type: string
        editIds:
          example: ["6421d1e887d1e2174b494492"]
          type: array
          items:
            type: string
        creationTime:
          example: "1975-08-19T23:15:30.000Z"
          description: Serialized JS Date object, see https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/toJSON
          type: string
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, possibly because it has expired.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
    InvalidError:
      description: Parameters don't match the schema (validated with [YUP](https://www.npmjs.com/package/yup)) or an internal server error was encountered.
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                description: An optional message describing the error
                example: "field body.recipientName missing"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT